move announcementconfig code to announcementconfigcontainer
